+++ lib/td_bg/comments/comments.ex
@@ -3,10 +3,12 @@ defmodule TdBg.Comments do
   The Comments context.
   """
 
-  import Ecto.Query, warn: false
-  alias TdBg.Repo
+  import Ecto.Query
 
+  alias Ecto.Multi
+  alias TdBg.Comments.Audit
   alias TdBg.Comments.Comment
+  alias TdBg.Repo
 
   @doc """
   Returns the list of comments.
@@ -21,7 +23,7 @@ defmodule TdBg.Comments do
     Repo.all(
       from(
         p in Comment,
-        order_by: [desc: :created_at]
+        order_by: [desc: :inserted_at]
       )
     )
   end
@@ -47,90 +49,69 @@ defmodule TdBg.Comments do
       from(
         p in Comment,
         where: ^dynamic,
-        order_by: [desc: :created_at]
+        order_by: [desc: :inserted_at]
       )
     )
   end
 
   @doc """
-  Gets a single comment.
-
-  Raises `Ecto.NoResultsError` if the Comment does not exist.
-
-  ## Examples
-
-      iex> get_comment!(123)
-      %Comment{}
-
-      iex> get_comment!(456)
-      ** (Ecto.NoResultsError)
-
-  """
-  def get_comment!(id), do: Repo.get!(Comment, id)
-
-  @doc """
-  Creates a comment.
+  Get a comment. Returns `{:ok, comment}` or `{:error, :not_found}` if the
+  comment doesn't exist.
 
   ## Examples
 
-      iex> create_comment(%{field: value})
+      iex> get_comment(123)
       {:ok, %Comment{}}
 
-      iex> create_comment(%{field: bad_value})
-      {:error, %Ecto.Changeset{}}
+      iex> get_comment(456)
+      {:error, :not_found}
 
   """
-  def create_comment(attrs \\ %{}) do
-    %Comment{}
-    |> Comment.changeset(attrs)
-    |> Repo.insert()
+  def get_comment(id) do
+    case Repo.get(Comment, id) do
+      nil -> {:error, :not_found}
+      comment -> {:ok, comment}
+    end
   end
 
   @doc """
-  Updates a comment.
+  Creates a comment and publishes the corresponding audit event.
 
   ## Examples
 
-      iex> update_comment(comment, %{field: new_value})
-      {:ok, %Comment{}}
+      iex> create_comment(%{field: value}, user)
+      {:ok, %{audit: "event_id", comment: %Comment{}}}
 
-      iex> update_comment(comment, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
+      iex> create_comment(%{field: bad_value}, user)
+      {:error, :comment, %Ecto.Changeset{}, %{}}
 
   """
-  def update_comment(%Comment{} = comment, attrs) do
-    comment
-    |> Comment.changeset(attrs)
-    |> Repo.update()
+  def create_comment(%{} = params, %{id: user_id}) do
+    changeset = Comment.changeset(params)
+
+    Multi.new()
+    |> Multi.insert(:comment, changeset)
+    |> Multi.run(:audit, Audit, :comment_created, [changeset, user_id])
+    |> Repo.transaction()
   end
 
   @doc """
-  Deletes a Comment.
+  Deletes a Comment and publishes the corresponding audit event.
 
   ## Examples
 
-      iex> delete_comment(comment)
-      {:ok, %Comment{}}
+      iex> delete_comment(comment, user)
+      {:ok, %{audit: "event_id", comment: %Comment{}}}
 
-      iex> delete_comment(comment)
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def delete_comment(%Comment{} = comment) do
-    Repo.delete(comment)
-  end
-
-  @doc """
-  Returns an `%Ecto.Changeset{}` for tracking comment changes.
-
-  ## Examples
-
-      iex> change_comment(comment)
-      %Ecto.Changeset{source: %Comment{}}
+      iex> delete_comment(comment, user)
+      {:error, :comment, %Ecto.Changeset{}, %{}}
 
   """
-  def change_comment(%Comment{} = comment) do
-    Comment.changeset(comment, %{})
+  def delete_comment(%Comment{} = comment, %{id: user_id}) do
+    Multi.new()
+    |> Multi.delete(:comment, comment)
+    |> Multi.run(:audit, Audit, :comment_deleted, [user_id])
+    |> Repo.transaction()
   end
 
   def get_comment_by_resource(resource_type, resource_id) do