+++ lib/td_bg_web/controllers/comment_controller.ex
@@ -4,8 +4,6 @@ defmodule TdBgWeb.CommentController do
   use PhoenixSwagger
 
   alias TdBg.Comments
-  alias TdBg.Comments.Comment
-  alias TdBg.Comments.Events
   alias TdBgWeb.SwaggerDefinitions
 
   action_fallback(TdBgWeb.FallbackController)
@@ -44,31 +42,21 @@ defmodule TdBgWeb.CommentController do
   end
 
   def create(conn, %{"comment" => comment_params}) do
-    current_user = conn.assigns[:current_user]
+    user = conn.assigns[:current_user]
 
     creation_attrs =
-      comment_params
-      |> Map.put("user", %{
-        "user_id" => current_user.id,
-        "full_name" => current_user.full_name,
-        "user_name" => current_user.user_name
+      Map.put(comment_params, "user", %{
+        "user_id" => user.id,
+        "full_name" => user.full_name,
+        "user_name" => user.user_name
       })
-      |> is_timestamp_informed?
 
-    with {:ok, %Comment{} = comment} <- Comments.create_comment(creation_attrs) do
-      Events.comment_created(comment, comment_params, conn.assigns[:current_user])
-
+    with {:ok, %{comment: comment}} <- Comments.create_comment(creation_attrs, user) do
       conn
       |> put_status(:created)
       |> put_resp_header("location", Routes.comment_path(conn, :show, comment))
       |> render("show.json", comment: comment)
     end
-  end
-
-  defp is_timestamp_informed?(comment_params) do
-    if Map.has_key?(comment_params, "created_at"),
-      do: comment_params,
-      else: Map.put(comment_params, "created_at", DateTime.utc_now())
   end
 
   swagger_path :show do
@@ -84,28 +72,7 @@ defmodule TdBgWeb.CommentController do
   end
 
   def show(conn, %{"id" => id}) do
-    comment = Comments.get_comment!(id)
-    render(conn, "show.json", comment: comment)
-  end
-
-  swagger_path :update do
-    description("Update Comments")
-    produces("application/json")
-
-    parameters do
-      id(:path, :integer, "Comment ID", required: true)
-      comment(:body, Schema.ref(:CommentUpdate), "Comment update attrs")
-    end
-
-    response(201, "OK", Schema.ref(:CommentResponse))
-    response(400, "Client Error")
-  end
-
-  def update(conn, %{"id" => id, "comment" => comment_params}) do
-    comment = Comments.get_comment!(id)
-
-    with {:ok, %Comment{} = comment} <- Comments.update_comment(comment, comment_params) do
-      Events.comment_updated(comment, comment_params, conn.assigns[:current_user])
+    with {:ok, comment} <- Comments.get_comment(id) do
       render(conn, "show.json", comment: comment)
     end
   end
@@ -123,10 +90,10 @@ defmodule TdBgWeb.CommentController do
   end
 
   def delete(conn, %{"id" => id}) do
-    comment = Comments.get_comment!(id)
+    user = conn.assigns[:current_user]
 
-    with {:ok, %Comment{}} <- Comments.delete_comment(comment) do
-      Events.comment_deleted(comment, conn.assigns[:current_user])
+    with {:ok, comment} <- Comments.get_comment(id),
+         {:ok, _} <- Comments.delete_comment(comment, user) do
       send_resp(conn, :no_content, "")
     end
   end